Modularize build with convention plugins and per-project files
