+++ build-logic/src/main/kotlin/scip.java-library.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+    id("scip.java-base")
+    `java-library`
+}
