+++ build.gradle.kts
@@ -1,471 +1,3 @@
-import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
-import com.google.protobuf.gradle.ProtobufExtension
-import com.google.protobuf.gradle.proto
-import com.vanniktech.maven.publish.MavenPublishBaseExtension
-import java.util.Properties
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    alias(libs.plugins.kotlin.jvm) apply false
-    alias(libs.plugins.protobuf) apply false
-    alias(libs.plugins.shadow) apply false
-    alias(libs.plugins.vanniktech.maven.publish) apply false
-}
-
-val javacInternalsFile = layout.projectDirectory.file("gradle/javac-internals.properties")
-val javacJvmOptions =
-    (
-        Properties()
-            .apply {
-                javacInternalsFile.asFile.inputStream().use { load(it) }
-            }
-            .getProperty("javac.jvmOptions")
-            ?: error("Missing javac.jvmOptions in gradle/javac-internals.properties")
-    )
-        .split(',')
-        .map { it.trim() }
-        .filter { it.isNotEmpty() }
-val catalog = libs
-
-allprojects {
-    group = "com.sourcegraph"
-    version = providers.gradleProperty("releaseVersion").orElse("0.0.0-SNAPSHOT").get()
-}
-
-subprojects {
-    // Several modules also have Bazel `BUILD` files. On the default macOS
-    // case-insensitive filesystem, Gradle's default `build/` directory collides
-    // with those files, so keep Gradle outputs under the already-ignored target/.
-    if (file("BUILD").isFile) {
-        layout.buildDirectory.set(layout.projectDirectory.dir("target/gradle"))
-    }
-
-    plugins.withType<JavaPlugin> {
-        tasks.withType<JavaCompile>().configureEach {
-            options.encoding = "UTF-8"
-            options.release.set(17)
-        }
-
-        tasks.withType<Javadoc>().configureEach {
-            options.encoding = "UTF-8"
-            (options as StandardJavadocDocletOptions).addStringOption("Xdoclint:none", "-quiet")
-        }
-
-        extensions.configure<TestingExtension> {
-            suites.named<JvmTestSuite>("test") {
-                useJUnitJupiter(catalog.versions.junit.jupiter)
-            }
-        }
-
-        tasks.withType<Test>().configureEach {
-            testLogging {
-                events("failed", "skipped")
-            }
-        }
-    }
-
-    plugins.withId("org.jetbrains.kotlin.jvm") {
-        tasks.withType<KotlinCompile>().configureEach {
-            compilerOptions.jvmTarget.set(JvmTarget.JVM_17)
-        }
-    }
-}
-
-fun Project.configureMavenPublishing() {
-    apply(plugin = "com.vanniktech.maven.publish")
-
-    extensions.configure<MavenPublishBaseExtension>("mavenPublishing") {
-        val repositoryUrl = "https://github.com/sourcegraph/scip-java"
-        publishToMavenCentral()
-        signAllPublications()
-
-        pom {
-            name.set(project.name)
-            description.set(project.description ?: project.name)
-            url.set(repositoryUrl)
-            licenses {
-                license {
-                    name.set("Apache-2.0")
-                    url.set("http://www.apache.org/licenses/LICENSE-2.0")
-                }
-            }
-            developers {
-                developer {
-                    id.set("sourcegraph")
-                    name.set("Sourcegraph")
-                }
-            }
-            scm {
-                connection.set("scm:git:$repositoryUrl.git")
-                developerConnection.set("scm:git:ssh://git@github.com/sourcegraph/scip-java.git")
-                url.set(repositoryUrl)
-            }
-        }
-    }
-}
-
-project(":scip-shared") {
-    apply(plugin = "java-library")
-    description = "Shared SCIP utilities used by scip-java compiler plugins"
-
-    dependencies {
-        "api"(catalog.scip.java.bindings)
-    }
-
-    configureMavenPublishing()
-}
-
-project(":scip-javac") {
-    apply(plugin = "java-library")
-    apply(plugin = "com.gradleup.shadow")
-    description = "A javac plugin to emit SCIP information"
-
-    dependencies {
-        "api"(project(":scip-shared"))
-    }
-
-    tasks.named<JavaCompile>("compileJava") {
-        options.isDebug = true
-        val emptyProcessorPath = layout.buildDirectory.dir("empty-processorpath")
-        options.annotationProcessorPath = files(emptyProcessorPath)
-        doFirst {
-            emptyProcessorPath.get().asFile.mkdirs()
-        }
-    }
-
-    tasks.named<Test>("test") {
-        jvmArgs(javacJvmOptions)
-    }
-
-    tasks.named<ShadowJar>("shadowJar") {
-        mergeServiceFiles()
-        exclude("javax/**", "com/sun/**", "sun/**", "META-INF/versions/9/module-info.class")
-        relocate("com.google", "com.sourcegraph.shaded.com.google")
-        relocate("google", "com.sourcegraph.shaded.google")
-        relocate("org.relaxng", "com.sourcegraph.shaded.relaxng")
-        relocate("com.sourcegraph", "com.sourcegraph.shaded.com.sourcegraph") {
-            exclude("com.sourcegraph.scip_javac.ScipPlugin")
-            exclude("com.sourcegraph.scip_javac.InjectScipOptions")
-        }
-    }
-
-    configureMavenPublishing()
-}
-
-project(":scip-kotlinc") {
-    apply(plugin = "java-library")
-    apply(plugin = "org.jetbrains.kotlin.jvm")
-    apply(plugin = "com.gradleup.shadow")
-    description = "A kotlinc plugin to emit SCIP information"
-
-    dependencies {
-        "implementation"(project(":scip-shared"))
-        "implementation"(catalog.scip.kotlin.bindings)
-        "compileOnly"(catalog.kotlin.stdlib)
-        "compileOnly"(catalog.kotlin.compiler.embeddable)
-
-        "testImplementation"(catalog.kotlin.compiler.embeddable)
-        "testImplementation"(catalog.kotlin.test)
-        "testImplementation"(catalog.kotlin.test.junit5)
-        "testImplementation"(catalog.kotlin.reflect)
-        "testImplementation"(catalog.kotest.assertions.core)
-        "testImplementation"(catalog.kctfork.core)
-    }
-
-    tasks.withType<KotlinCompile>().configureEach {
-        compilerOptions.freeCompilerArgs.add("-Xcontext-parameters")
-    }
-
-    tasks.named<Test>("test") {
-        maxHeapSize = "2g"
-    }
-
-    tasks.named<ShadowJar>("shadowJar") {
-        mergeServiceFiles()
-        relocate("com.intellij", "org.jetbrains.kotlin.com.intellij")
-    }
-
-    configureMavenPublishing()
-}
-
-project(":scip-gradle-plugin") {
-    apply(plugin = "java-library")
-    apply(plugin = "com.gradleup.shadow")
-
-    dependencies {
-        "compileOnly"(catalog.gradle.api)
-        "compileOnly"(catalog.gradle.test.kit)
-    }
-}
-
-project(":scip-maven-plugin") {
-    apply(plugin = "java-library")
-    description = "A Maven plugin that exports dependency metadata for scip-java"
-
-    dependencies {
-        "implementation"(catalog.maven.plugin.api)
-        "implementation"(catalog.maven.project)
-        "compileOnly"(catalog.maven.plugin.annotations)
-    }
-
-    tasks.named<ProcessResources>("processResources") {
-        exclude("META-INF/maven/plugin.template.xml")
-        from("src/main/resources/META-INF/maven/plugin.template.xml") {
-            into("META-INF/maven")
-            rename { "plugin.xml" }
-            filter { line: String -> line.replace("@VERSION@", project.version.toString()) }
-        }
-        inputs.property("version", project.version.toString())
-    }
-
-    configureMavenPublishing()
-}
-
-project(":scip-aggregator") {
-    apply(plugin = "java-library")
-    apply(plugin = "com.google.protobuf")
-    description = "Aggregates compiler-plugin SCIP shards into a single SCIP index"
-
-    dependencies {
-        "api"(catalog.scip.java.bindings)
-        "implementation"(project(":scip-shared"))
-    }
-
-    extensions.configure<ProtobufExtension>("protobuf") {
-        val protobufVersion = catalog.versions.protobuf.asProvider().get()
-        protoc {
-            artifact = "com.google.protobuf:protoc:$protobufVersion"
-        }
-    }
-
-    extensions.configure<SourceSetContainer> {
-        named("main") {
-            proto {
-                srcDir("src/main/protobuf")
-            }
-        }
-    }
-
-    configureMavenPublishing()
-}
-
-project(":scip-java") {
-    apply(plugin = "org.jetbrains.kotlin.jvm")
-    apply(plugin = "application")
-    description = "Java and Kotlin indexer for SCIP"
-
-    dependencies {
-        "implementation"(project(":scip-aggregator"))
-        "implementation"(catalog.clikt.jvm)
-        "implementation"(catalog.kotlin.stdlib)
-        "implementation"(catalog.kotlin.compiler.embeddable)
-        "implementation"(catalog.kotlin.scripting.common)
-        "implementation"(catalog.kotlin.scripting.jvm)
-        "implementation"(catalog.kotlin.scripting.dependencies)
-        "implementation"(catalog.kotlin.scripting.dependencies.maven)
-        "implementation"(catalog.kotlinx.serialization.json.jvm)
-
-        "testImplementation"(catalog.kotlin.test)
-        "testImplementation"(catalog.kotlin.test.junit5)
-    }
-
-    tasks.named<Test>("test") {
-        jvmArgs(javacJvmOptions)
-        systemProperty("scip.jdk.version", "11")
-    }
-
-    extensions.configure<JavaApplication> {
-        mainClass.set("com.sourcegraph.scip_java.ScipJava")
-    }
-
-    val generateEmbeddedResources = tasks.register<Sync>("generateEmbeddedResources") {
-        val javacShadow = project(":scip-javac").tasks.named<ShadowJar>("shadowJar")
-        val gradleShadow = project(":scip-gradle-plugin").tasks.named<ShadowJar>("shadowJar")
-        val kotlincShadow = project(":scip-kotlinc").tasks.named<ShadowJar>("shadowJar")
-
-        from(javacShadow.flatMap { it.archiveFile }) {
-            rename { "scip-plugin.jar" }
-        }
-        from(gradleShadow.flatMap { it.archiveFile }) {
-            rename { "gradle-plugin.jar" }
-        }
-        from(kotlincShadow.flatMap { it.archiveFile }) {
-            rename { "scip-kotlinc.jar" }
-        }
-        into(layout.buildDirectory.dir("generated/resources/embedded"))
-    }
-
-    tasks.named<ProcessResources>("processResources") {
-        from(javacInternalsFile)
-        from(generateEmbeddedResources)
-    }
-
-    tasks.named<Jar>("jar") {
-        manifest {
-            attributes("Implementation-Version" to project.version)
-        }
-    }
-
-    val generateDistributionVersion = tasks.register("generateDistributionVersion") {
-        val output = layout.buildDirectory.file("generated/distribution/VERSION")
-        val version = project.version.toString()
-        inputs.property("version", version)
-        outputs.file(output)
-        doLast {
-            val file = output.get().asFile
-            file.parentFile.mkdirs()
-            file.writeText("version:=$version\n")
-        }
-    }
-
-    // Emit the javac --add-exports flags as a ready-to-use, space-separated
-    // string so the Docker wrapper script can read them with a plain `cat`,
-    // keeping gradle/javac-internals.properties the single source of truth.
-    val generateJavacJvmOptions = tasks.register("generateJavacJvmOptions") {
-        val output = layout.buildDirectory.file("generated/distribution/javac-jvm-options")
-        val options = javacJvmOptions.joinToString(" ")
-        inputs.property("options", options)
-        outputs.file(output)
-        doLast {
-            val file = output.get().asFile
-            file.parentFile.mkdirs()
-            file.writeText("$options\n")
-        }
-    }
-
-    extensions.configure<DistributionContainer> {
-        named("main") {
-            contents {
-                from(generateDistributionVersion)
-                from(generateJavacJvmOptions)
-            }
-        }
-    }
-
-    configureMavenPublishing()
-}
-
-project(":scip-snapshots-java-common") {
-    apply(plugin = "java-library")
-
-    dependencies {
-        "compileOnly"(catalog.lombok)
-        "annotationProcessor"(catalog.lombok)
-    }
-
-    val annotationProcessorClasspath = configurations.named("annotationProcessor")
-
-    tasks.named<JavaCompile>("compileJava") {
-        val javacShadow = project(":scip-javac").tasks.named<ShadowJar>("shadowJar")
-        val scipTargetroot = layout.buildDirectory.dir("scip-targetroot")
-        dependsOn(javacShadow)
-        classpath = classpath.plus(files(javacShadow.flatMap { it.archiveFile }))
-        options.annotationProcessorPath =
-            annotationProcessorClasspath.get().plus(files(javacShadow.flatMap { it.archiveFile }))
-        outputs.dir(scipTargetroot)
-        options.isFork = true
-        options.forkOptions.jvmArgs = (options.forkOptions.jvmArgs ?: emptyList()) + javacJvmOptions
-        options.compilerArgs.add(
-            "-Xplugin:scip -text:on -verbose -sourceroot:${rootProject.rootDir.absolutePath} " +
-                "-targetroot:${scipTargetroot.get().asFile.absolutePath} -randomtimestamp=${System.nanoTime()}"
-        )
-        doFirst {
-            scipTargetroot.get().asFile.deleteRecursively()
-            scipTargetroot.get().asFile.mkdirs()
-        }
-    }
-}
-
-project(":scip-snapshots-kotlin-common") {
-    apply(plugin = "java-library")
-    apply(plugin = "org.jetbrains.kotlin.jvm")
-
-    dependencies {
-        "implementation"(catalog.kotlin.stdlib)
-    }
-
-    val scipTargetroot = layout.buildDirectory.dir("scip-targetroot")
-    val javacShadow = project(":scip-javac").tasks.named<ShadowJar>("shadowJar")
-    val kotlincShadow = project(":scip-kotlinc").tasks.named<ShadowJar>("shadowJar")
-
-    tasks.named<KotlinCompile>("compileKotlin") {
-        dependsOn(kotlincShadow)
-        inputs.file(kotlincShadow.flatMap { it.archiveFile })
-        outputs.dir(scipTargetroot)
-        compilerOptions.freeCompilerArgs.addAll(
-            "-Xplugin=${kotlincShadow.flatMap { it.archiveFile }.get().asFile.absolutePath}",
-            "-P",
-            "plugin:scip-kotlinc:sourceroot=${rootProject.rootDir.absolutePath}",
-            "-P",
-            "plugin:scip-kotlinc:targetroot=${scipTargetroot.get().asFile.absolutePath}",
-        )
-        doFirst {
-            scipTargetroot.get().asFile.deleteRecursively()
-            scipTargetroot.get().asFile.mkdirs()
-        }
-    }
-
-    tasks.named<JavaCompile>("compileJava") {
-        dependsOn(javacShadow)
-        classpath = classpath.plus(files(javacShadow.flatMap { it.archiveFile }))
-        options.annotationProcessorPath = files(javacShadow.flatMap { it.archiveFile })
-        outputs.dir(scipTargetroot)
-        options.isFork = true
-        options.forkOptions.jvmArgs = (options.forkOptions.jvmArgs ?: emptyList()) + javacJvmOptions
-        options.compilerArgs.add(
-            "-Xplugin:scip -sourceroot:${rootProject.rootDir.absolutePath} " +
-                "-targetroot:${scipTargetroot.get().asFile.absolutePath}"
-        )
-    }
-}
-
-project(":scip-snapshots") {
-    apply(plugin = "java-library")
-
-    dependencies {
-        "implementation"(project(":scip-java"))
-        "implementation"(catalog.scip.java.bindings)
-    }
-
-    val javaCase = project(":scip-snapshots-java-common")
-    val kotlinCase = project(":scip-snapshots-kotlin-common")
-    val javaCaseClasses = javaCase.tasks.named("classes")
-    val kotlinCaseClasses = kotlinCase.tasks.named("classes")
-    val javaTargetroot = javaCase.layout.buildDirectory.dir("scip-targetroot")
-    val kotlinTargetroot = kotlinCase.layout.buildDirectory.dir("scip-targetroot")
-    val snapshotProperties =
-        mapOf(
-            "snapshot.sourceroot" to rootProject.rootDir.absolutePath,
-            "snapshot.cases" to "java-common,kotlin-common",
-            "snapshot.case.java-common.expectDir" to
-                rootProject.layout.projectDirectory.dir("scip-snapshots/expected/java/common").asFile.absolutePath,
-            "snapshot.case.java-common.targetroot" to javaTargetroot.get().asFile.absolutePath,
-            "snapshot.case.kotlin-common.expectDir" to
-                rootProject.layout.projectDirectory.dir("scip-snapshots/expected/kotlin/common").asFile.absolutePath,
-            "snapshot.case.kotlin-common.targetroot" to kotlinTargetroot.get().asFile.absolutePath,
-            "snapshot.case.kotlin-common.aggregateNoEmitInverseRelationships" to "true",
-            "scip.jdk.version" to "17",
-        )
-
-    tasks.named<Test>("test") {
-        dependsOn(javaCaseClasses, kotlinCaseClasses)
-        jvmArgs(javacJvmOptions)
-        systemProperties(snapshotProperties)
-    }
-
-    tasks.register<JavaExec>("saveSnapshots") {
-        group = "verification"
-        description = "Regenerates Java and Kotlin SCIP snapshot goldens."
-        dependsOn(tasks.named("classes"), javaCaseClasses, kotlinCaseClasses)
-        val sourceSets = project.extensions.getByType<SourceSetContainer>()
-        classpath = sourceSets.named("main").get().runtimeClasspath
-        mainClass.set("tests.SaveSnapshots")
-        jvmArgs(javacJvmOptions)
-        systemProperties(snapshotProperties)
-    }
-}
-
 tasks.register("saveSnapshots") {
     group = "verification"
     description = "Regenerates all SCIP snapshot goldens."
