+++ scip-gradle-plugin/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("scip.java-library")
+    id("scip.shadow-producer")
+}
+
+dependencies {
+    compileOnly(libs.gradle.api)
+    compileOnly(libs.gradle.test.kit)
+}
