+++ scip-shared/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    id("scip.java-library")
+    id("scip.maven-publish")
+}
+
+description = "Shared SCIP utilities used by scip-java compiler plugins"
+
+dependencies {
+    api(libs.scip.java.bindings)
+}
